refactor: PR 반영 및 로고 눌렀을때 메인으로 이동
