+++ src/containers/common/Header/Header.tsx
@@ -1,11 +1,9 @@
-import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet";
 import { useIsMobile } from "@/hooks/useIsMobile";
 import { cn } from "@/libs/utils";
-import { List, CaretDown } from "@phosphor-icons/react";
+import { List} from "@phosphor-icons/react";
 import { Link } from "react-router-dom";
-import DropDownMenu from "@/components/DropDownMenu";
-import { menuItems } from "@/components/DropDownMenu/const";
-import { dataPath, State } from "./const/state";
+import DropDownMenu from "@/containers/common/Header/component/DropDownMenu";
+import { dataPath, menuItems } from "@/containers/common/Header/const/pathData";
 import { Logo } from "@/components/Logo/Logo";
 import {
   NavigationMenu,
@@ -15,19 +13,11 @@ import {
   NavigationMenuContent,
 } from "@/components/ui/navigation-menu";
 import { getStyles } from "./const/style";
-import { ReactNode, useState } from "react";
+import { HeaderSheet } from "./component/HeaderSheet";
+import { State } from "./const/state";
 
 interface HeaderProps {
   state?: State;
-}
-
-interface HeaderSheetProps {
-  trigger: ReactNode;
-  sheetBorderColor: string;
-  bgColor: string;
-  textColor: string;
-  sheetIconColor: string;
-  sheetItemsColor: string;
 }
 
 export function Header({ state = State.Onboarding }: HeaderProps) {
@@ -35,21 +25,17 @@ export function Header({ state = State.Onboarding }: HeaderProps) {
   const styles = getStyles(state);
 
   return (
-    <div
-      className={cn(
-        "fixed top-0 w-[100vw] h-[60px] sm:h-[50px] xs:h-[50px] z-60",
+    // prettier ignore
+    <div className={cn(
+        "fixed top-0 w-[100vw] h-[60px] sm:h-[50px] xs:h-[50px]",
         isMobile ? "flex justify-start" : "flex justify-between",
-        styles.bgColor
-      )}
-      style={{ zIndex: 100 }}
-    >
+        styles.bgColor,)}
+      style = {{zIndex: 100}}>
       {isMobile && (
         <HeaderSheet
-          trigger={
-            <div className={cn(styles.headerItemStyle, "px-3 text-base")}>
+          trigger={<div className={cn(styles.headerItemStyle, "px-3 text-base")}>
               <List size={28} />
-            </div>
-          }
+            </div>}
           sheetBorderColor={styles.sheetBorderColor}
           bgColor={styles.bgColor}
           textColor={styles.textColor}
@@ -57,7 +43,7 @@ export function Header({ state = State.Onboarding }: HeaderProps) {
           sheetItemsColor={styles.sheetItemsColor}
         />
       )}
-      <div className="flex">
+      <Link to='/' className="flex">
         <div className={cn(styles.headerItemStyle, { "px-0.5": isMobile })}>
           <div className="flex items-center gap-4">
             <Logo size={isMobile ? "23px" : "46px"} fill={styles.fillColor} />
@@ -72,20 +58,18 @@ export function Header({ state = State.Onboarding }: HeaderProps) {
           </div>
         </div>
         {!isMobile && renderNavigationMenu()}
-      </div>
+      </Link>
       {!isMobile && <div className="flex pr-4">{renderAuthButtons()}</div>}
     </div>
   );
 
   function renderNavigationMenu() {
     return (
+      // pretier ignore
       <NavigationMenu className="h-full">
         <NavigationMenuList className="h-full">
           {Object.entries(menuItems).map(([category, items]) => (
-            <NavigationMenuItem
-              key={category}
-              className="relative h-full text-[20px] min-w-fit"
-            >
+            <NavigationMenuItem key={category} className="relative h-full text-[20px] min-w-fit">
               <NavigationMenuTrigger className={cn(styles.headerItemStyle)}>
                 {category}
               </NavigationMenuTrigger>
@@ -100,10 +84,7 @@ export function Header({ state = State.Onboarding }: HeaderProps) {
             </NavigationMenuItem>
           ))}
           <NavigationMenuItem className="relative h-full text-[20px] min-w-fit">
-            <NavigationMenuTrigger
-              isData={true}
-              className={cn(styles.headerItemStyle)}
-            >
+            <NavigationMenuTrigger isData={true} className={cn(styles.headerItemStyle)}>
               <Link to={dataPath}>자료집</Link>
             </NavigationMenuTrigger>
           </NavigationMenuItem>
@@ -112,93 +93,17 @@ export function Header({ state = State.Onboarding }: HeaderProps) {
     );
   }
 
-  function HeaderSheet({
-    trigger,
-    sheetBorderColor,
-    bgColor,
-    textColor,
-    sheetIconColor,
-    sheetItemsColor,
-  }: HeaderSheetProps) {
-    const [expandedCategory, setExpandedCategory] = useState<string | null>(
-      null
-    );
-
-    const toggleCategory = (category: string) => {
-      setExpandedCategory((prevCategory) =>
-        prevCategory === category ? null : category
-      );
-    };
-
-    return (
-      <Sheet>
-        <SheetTrigger asChild>{trigger}</SheetTrigger>
-        <SheetContent
-          className={`top-[50px] left-0 border-0 outline-none flex items-start justify-start w-[280px] px-0 py-0 ${bgColor}`}
-        >
-          <div className="flex flex-col w-full text-lg font-semibold">
-            {Object.entries(menuItems).map(([category, items], index) => (
-              <div key={index} className="w-full">
-                <div
-                  className={`w-full h-[64px] flex flex-row items-center justify-between pl-10 border-b ${sheetBorderColor} cursor-pointer`}
-                  onClick={() => toggleCategory(category)}
-                >
-                  <div className={`flex-1 flex items-center ${textColor}`}>
-                    {category}
-                  </div>
-                  <CaretDown className={`${sheetIconColor}`} size={20} />
-                  <div className="w-3"></div>
-                </div>
-                {expandedCategory === category && (
-                  <div
-                    className={`flex flex-col flex-center justify-center py-4 ${bgColor} border-b ${sheetBorderColor}`}
-                  >
-                    {items.map((item) => (
-                      <Link
-                        key={item.path}
-                        to={item.path}
-                        className={`flex items-center h-[32px] px-4 pl-12 font-medium text-base ${sheetItemsColor}`}
-                      >
-                        {item.name}
-                      </Link>
-                    ))}
-                  </div>
-                )}
-              </div>
-            ))}
-            <Link
-              to={dataPath}
-              className={`h-[64px] flex items-center pl-10 border-b ${sheetBorderColor} ${textColor}`}
-            >
-              자료집
-            </Link>
-          </div>
-        </SheetContent>
-      </Sheet>
-    );
-  }
-
   function renderAuthButtons() {
     if (state === State.Login) {
       return (
-        <div
-          className={cn(
-            styles.headerItemStyle,
-            "px-10 text-base text-[20px] min-w-fit"
-          )}
-        >
+        <div className={cn(styles.headerItemStyle,"px-10 text-base text-[20px] min-w-fit")}>
           내정보
         </div>
       );
     }
     if (state === State.Logout) {
       return (
-        <div
-          className={cn(
-            styles.headerItemStyle,
-            "px-10 text-base text-[20px] min-w-fit"
-          )}
-        >
+        <div className={cn(styles.headerItemStyle,"px-10 text-base text-[20px] min-w-fit")}>
           로그인
         </div>
       );
